requirements.txt
boto3
botocore
langdetect
PyPDF2
python-docx
pillow
pytesseract
requests
